Extract common methods out of QlIntelBaseCC
